修复 Windows x64 编译错误和 macOS 运行器配置
- v8handler_impl.cc: 添加 static_cast<int> 转换解决 size_t 到 int 的警告 (C4267)
- build.yml: 将 macos-13 改为 macos-14 (GitHub 已弃用 macos-13 运行器)

+++ src/v8handler_impl.cc
@@ -196,9 +196,9 @@ CefRefPtr<CefV8Value> V8HandlerImpl::ConvertCefValueToV8(
       return CefV8Value::CreateString(value->GetString());
     case VTYPE_LIST: {
       CefRefPtr<CefListValue> list = value->GetList();
-      CefRefPtr<CefV8Value> array = CefV8Value::CreateArray(list->GetSize());
+      CefRefPtr<CefV8Value> array = CefV8Value::CreateArray(static_cast<int>(list->GetSize()));
       for (size_t i = 0; i < list->GetSize(); i++) {
-        array->SetValue(i, ConvertCefValueToV8(list->GetValue(i)));
+        array->SetValue(static_cast<int>(i), ConvertCefValueToV8(list->GetValue(i)));
       }
       return array;
     }
